-Made the memory map use ushort arrays because the Intellivision is 16-bit. -Fixed JMP disassembly; I need to return on an invalid opcode because I was breaking out of the inner switch statement, not both that and the outer one.

+++ BizHawk.Emulation/CPUs/CP1610/CP1610.cs
@@ -15,7 +15,7 @@ namespace BizHawk.Emulation.CPUs.CP1610
 		public int TotalExecutedCycles;
 		public int PendingCycles;
 
-		public Func<ushort, byte> ReadMemory;
-		public Action<ushort, byte> WriteMemory;
+		public Func<ushort, ushort> ReadMemory;
+		public Action<ushort, ushort> WriteMemory;
 	}
 }
